Implement optimal raw comparator choosing
If an index key has one of special structures then for
tuples of this index optimized comparator can be used.

The choosing of optimal comparator was implemented for
not raw comparators (that use struct tuple).

This commit reimplements optimal comparator choosing
so that now a raw comparator is choosed and it can be used
in vinyl.c.

See #1572