vinyl: coalesce small ranges
After a range was successfully compacted or dumped, consider coalescing
it with its neighbors. Only coalesce two adjacent ranges if the total
amount of disk space occupied by them is < range_size / 2, to avoid
split/coalesce oscillations. Coalescence is done by linking runs and
mems of the ranges together and deleting one of the ranges while
updating boundaries of another as necessary, i.e. it doesn't require any
disk writes. The change is committed to disk only on the subsequent
compaction. That means, we cannot dump the resulting range before it is
compacted.

Closes #1735