vinyl: create in-memory indexes on demand
When we move vinyl metadata to memtx, we won't be able to create
in-memory indexes on dump, because if we created a run record in the
metadata table during checkpoint it wouldn't be added to the snapshot,
making recovery impossible. So in this patch we make vinyl create
in-memory indexes on-demand, i.e. when the first transaction is
processed.